+++ dao/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog 刀 DAO
 # Day Ahead Optimizer
+## [V2024.11.0.dev_a]
+- Scheduling of heatpump can be set via a ha-entity
+- Scheduling of heatpump can be achieved in three ways:
+  - on/off
+  - heat curve adjustment
+  - calculated power
+- When setting a state of an entity failed then an error message is written in the log 
+  (name of the entity, new failed value)
+
+## [V2024.11.0]
+- Fixed a db-error when getting Tibber-data with the default sqlite-db
 
 ## [V2024.10.6]
 - Fixed a few errors in the graphics for users with more than one battery